Move some code around to prepare for merging PageRecycler fixes.
