+++ src/llfs/page_recycler_state.cpp
@@ -40,6 +40,8 @@ PageRecycler::State::State(const boost::uuids::uuid& uuid,
   BATT_CHECK_EQ(2, options.refresh_factor) << "implement generic refresh factors";
 }
 
+//==#==========+==+=+=++=+++++++++++-+-+--+----- --- -- -  -  -   -
+//
 void PageRecycler::State::bulk_load(Slice<const PageToRecycle> pages)
 {
   Optional<slot_offset_type> prev_slot;
